+++ src/js/visual/configurator-prc.js
@@ -1546,6 +1546,12 @@ ripe.ConfiguratorPrc.prototype._registerHandlers = function() {
         event.stopPropagation();
     });
 
+    area.addEventListener("mouseleave", function(event) {
+        // in case the mouse leaves the area then the
+        // part highlight must be removed
+        self.lowlight();
+    });
+
     area.addEventListener("mousemove", function(event) {
         const preloading = self.element.classList.contains("preloading");
         const animating = self.element.classList.contains("animating");